Move all files into packages
