+++ src/main/java/Main.java
@@ -1,3 +1,5 @@
+import ari.Ari;
+
 public class Main {
     public static void main(String[] args) {
         Ari ari = new Ari("data/ari.txt");
