+++ src/main/java/ari/command/ByeCommand.java
@@ -1,3 +1,5 @@
+package ari.command;
+
 public class ByeCommand extends Command {
     public static final String COMMAND_WORD = "bye";
