+++ src/main/java/ari/command/IncorrectCommand.java
@@ -1,3 +1,5 @@
+package ari.command;
+
 public class IncorrectCommand extends Command {
     private String message;
