+++ src/main/java/ari/command/TodoCommand.java
@@ -1,3 +1,7 @@
+package ari.command;
+
+import ari.tasks.ToDoTask;
+
 public class TodoCommand extends AddCommand {
     public static final String COMMAND_WORD = "todo";
