+++ src/main/java/ari/command/UnmarkCommand.java
@@ -1,3 +1,5 @@
+package ari.command;
+
 public class UnmarkCommand extends Command {
     public static final String COMMAND_WORD = "unmark";
