+++ src/main/java/ari/exception/DukeException.java
@@ -1,3 +1,5 @@
+package ari.exception;
+
 public class DukeException extends IllegalArgumentException {
     public DukeException() {
     }
