+++ src/main/java/ari/exception/EmptyCommandException.java
@@ -1,3 +1,5 @@
+package ari.exception;
+
 public class EmptyCommandException extends DukeException {
     public EmptyCommandException() {
         super("Sorry Master, I did not catch that. Could you repeat?");
