+++ src/main/java/ari/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package ari.exception;
+
 public class InvalidCommandException extends DukeException {
 
     public InvalidCommandException(String message) {
