+++ src/main/java/ari/ui/Ui.java
@@ -1,3 +1,5 @@
+package ari.ui;
+
 import java.util.Scanner;
 
 public class Ui { // deals with interactions with the user
